get-papers/pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "CLI tool to fetch PubMed papers with pharma/biotech authors."
authors = ["Rakshitha N <nampallyrakshitha@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "get_papers" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
biopython = "^1.83"


[tool.poetry.scripts]
get-papers-list = "get_papers.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

